Avinya_Assistant/requirements.txt
SpeechRecognition
pyttsx3
pyaudio
google-generativeai
python-dotenv
pywhatkit
